Consolidate connections controller/module files
